test: added tests for face anon stats

+++ ampc-anon-stats/src/server/config/mod.rs
@@ -53,15 +53,15 @@ pub struct AnonStatsServerConfig {
 
     #[serde(default = "default_face_threshold_start")]
     /// Lower threshold for face anon stats histogram bins.
-    pub face_threshold_start: u16,
+    pub face_threshold_start: i16,
 
     #[serde(default = "default_face_threshold_end")]
     /// Upper threshold for face anon stats histogram bins.
-    pub face_threshold_end: u16,
+    pub face_threshold_end: i16,
 
     #[serde(default = "default_face_threshold_step")]
     /// Bin size for face anon stats histogram.
-    pub face_threshold_step: u16,
+    pub face_threshold_step: usize,
 
     #[serde(default = "default_max_sync_failures_before_reset")]
     /// Number of consecutive sync mismatches before clearing the local queue for an origin.
@@ -85,15 +85,15 @@ pub struct AnonStatsServerConfig {
     pub shutdown_last_results_sync_timeout_secs: u64,
 }
 
-fn default_face_threshold_start() -> u16 {
+fn default_face_threshold_start() -> i16 {
     0
 }
 
-fn default_face_threshold_end() -> u16 {
+fn default_face_threshold_end() -> i16 {
     5000
 }
 
-fn default_face_threshold_step() -> u16 {
+fn default_face_threshold_step() -> usize {
     100
 }
